actions: Split out zerver.actions.message_edit.
(cherry picked from commit eda000899b35c69e475fb311646cbd1201f9d319)
