+++ static/js/message_edit.js
@@ -401,7 +401,7 @@ function edit_message($row, raw_content) {
     // been able to click it at the time the mouse entered the message_row. Also
     // a buffer in case their computer is slow, or stalled for a second, etc
     // If you change this number also change edit_limit_buffer in
-    // zerver.lib.actions.check_update_message
+    // zerver.actions.message_edit.check_update_message
     const seconds_left_buffer = 5;
     const editability = get_editability(message, seconds_left_buffer);
     const max_file_upload_size = page_params.max_file_upload_size_mib;
@@ -534,7 +534,7 @@ function edit_message($row, raw_content) {
     ) {
         // Give them at least 10 seconds.
         // If you change this number also change edit_limit_buffer in
-        // zerver.lib.actions.check_update_message
+        // zerver.actions.message_edit.check_update_message
         const min_seconds_to_edit = 10;
         let seconds_left =
             page_params.realm_message_content_edit_limit_seconds +
