+++ zerver/tests/test_example.py
@@ -494,7 +494,7 @@ class TestMocking(ZulipTestCase):
         # that is beyond the limit.
         #
         # Notice how mock.patch() is used here to do exactly the above mentioned.
-        # mock.patch() here makes any calls to `timezone_now` in `zerver.lib.actions`
+        # mock.patch() here makes any calls to `timezone_now` in `zerver.actions.message_edit`
         # to return the value passed to `return_value` in the its context.
         # You can also use mock.patch() as a decorator depending on the
         # requirements. Read more at the documentation link provided above.
@@ -504,7 +504,7 @@ class TestMocking(ZulipTestCase):
         )  # There's a buffer time applied to the limit, hence the extra 100s.
 
         with mock.patch(
-            "zerver.lib.actions.timezone_now",
+            "zerver.actions.message_edit.timezone_now",
             return_value=time_beyond_edit_limit,
         ):
             result = self.client_patch(
