+++ zerver/tests/test_link_embed.py
@@ -389,7 +389,7 @@ class PreviewTestCase(ZulipTestCase):
         url = "http://test.org/"
         self.create_mock_response(url)
 
-        with mock_queue_publish("zerver.lib.actions.queue_json_publish") as patched:
+        with mock_queue_publish("zerver.actions.message_edit.queue_json_publish") as patched:
             result = self.client_patch(
                 "/json/messages/" + str(msg_id),
                 {
@@ -509,7 +509,7 @@ class PreviewTestCase(ZulipTestCase):
             )
 
         with mock_queue_publish(
-            "zerver.lib.actions.queue_json_publish", wraps=wrapped_queue_json_publish
+            "zerver.actions.message_edit.queue_json_publish", wraps=wrapped_queue_json_publish
         ):
             result = self.client_patch(
                 "/json/messages/" + str(msg_id),
