+++ zerver/worker/queue_processors.py
@@ -47,13 +47,13 @@ from sentry_sdk import add_breadcrumb, configure_scope
 from zulip_bots.lib import extract_query_without_mention
 
 from zerver.actions.invites import do_send_confirmation_email
+from zerver.actions.message_edit import do_update_embedded_data
 from zerver.actions.message_flags import do_mark_stream_messages_as_read
 from zerver.actions.message_send import internal_send_private_message, render_incoming_message
 from zerver.actions.presence import do_update_user_presence
 from zerver.actions.realm_export import notify_realm_export
 from zerver.actions.user_activity import do_update_user_activity, do_update_user_activity_interval
 from zerver.context_processors import common_context
-from zerver.lib.actions import do_update_embedded_data
 from zerver.lib.bot_lib import EmbeddedBotHandler, EmbeddedBotQuitException, get_bot_handler
 from zerver.lib.context_managers import lockfile
 from zerver.lib.db import reset_queries
@@ -865,7 +865,7 @@ class FetchLinksEmbedData(QueueProcessingWorker):
 
         message = Message.objects.get(id=event["message_id"])
         # If the message changed, we will run this task after updating the message
-        # in zerver.lib.actions.check_update_message
+        # in zerver.actions.message_edit.check_update_message
         if message.content != event["message_content"]:
             return
         if message.content is not None:
